convert float to int

+++ math/0x01-calculus/9-sum_total.py
@@ -4,4 +4,4 @@
 
 def summation_i_squared(n):
     """ Calculates the summation of i^2 from i = 1 to n """
-    return (n * (n + 1) * (2 * n + 1)) / 6
+    return int((n * (n + 1) * (2 * n + 1)) / 6)
